2:09

+++ my_lib/ft_memnjoin.c
@@ -6,13 +6,13 @@
 /*   By: ksharlen <ksharlen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/13 15:05:30 by ksharlen          #+#    #+#             */
-/*   Updated: 2019/05/13 23:51:48 by ksharlen         ###   ########.fr       */
+/*   Updated: 2019/05/15 01:32:02 by ksharlen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	*ft_memnjoin(const void *s1, const void *s2, int num1, int num2)
+void	*ft_memnjoin(const void *s1, const void *s2, size_t num1, size_t num2)
 {
 	void	*p_ret;
